[43475101] Use the correct variable naming conventions
Changed some variable names from camelCase to
snake_case and renamed some variables.
